requirements.txt
asgiref==3.8.1
attrs==24.2.0
cattrs==24.1.2
dj-database-url==2.2.0
Django==5.1.1
django-debug-toolbar==4.4.6
django-extensions==3.2.3
django-filter==24.3
django-login-required-middleware==0.9.0
django-schema-graph==3.1.0
djangorestframework==3.15.2
Markdown==3.7
pillow==10.4.0
psycopg==3.2.2
python-dotenv==1.0.1
sqlparse==0.5.1
typing_extensions==4.12.2
